requirements.txt
Django==2.2.7
django-solo==1.1.3
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
virtualenv==16.7.5
